+++ CHANGELOG.md
@@ -11,11 +11,16 @@ Versioning](http://semver.org/spec/v2.0.0.html).
 
 ### Added
 - Added per resource counts to tessen data collection.
+- Added event processing counts to tessen data collection.
 - Added ability to query for `Handlers` (individual and collections) from the GraphQL query endpoint.
 
 ### Changed
 - GraphQL fields that previously accepted a JS filter have been deprecated and
   replaced with a simpler syntax.
+
+### Fixed
+- Only bury switchsets of checks that no longer have a TTL, in order to reduce
+the number of write operations made to etcd.
 
 ## [5.7.0] - 2019-05-09